Actualizacion de contraseña e informacion de perfil

+++ database/migrations/2022_04_14_220103_add_fields_phone_birth_gender_to_users_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->integer('phone')->after('avatar')->nullable();
+            $table->bigInteger('phone')->after('avatar')->nullable();
             $table->date('birthday')->after('phone')->nullable();
             $table->integer('gender')->after('birthday')->nullable();
         });
